Trace client-side certificates
Implement tracing of peer certificates (client-side certificates) while
the verification process is in progress. This allows to spot mistakes of
the remote party. Currently a certificate that is not in the trusted
list just receives "bad certificate". Also the Cherokee side is far from
informal. This patch addresses it.

The originial change contained the function SSL_get_peer_certificate.
It might be an interesting new rule, to validate on a specific peer
certificate or specific properies thereof.